% Generated by roxygen2 (4.0.1): do not edit by hand
\name{medd}
\alias{medd}
\title{medd}
\usage{
medd(x)
}
\arguments{
\item{x}{data frame entanyl_iv IV fentaNYL total}
}
\value{
a data frame with the original columns, new morphine equivalent columns
and the total morphine equivalent for each row
}
\description{
converts the total dosages of various opioid medications and routes
into a Morphine Equivalent Daily Dose (MEDD)
}
\examples{
library(dplyr)
data <- data.frame(
  names=c("identity", "john", "jane", "jane"),
  pod=c(1,1,1,2),
  hydrocodone_po=c(1,5,1,1),
  morphine_iv=c(1,0,1,1),
  morphine_po=c(1,0,1,1),
  nalbuphine_iv=c(1,0,10,10),
  hydromorphone_iv = c(1,1,2,2),
  fentanyl_iv = c(1,2,2,0),
  fentanyl_td = c(1,1,1,1),
  hydromorphone_po = c(1,NA, 2, 2),
  methadone_iv = c(1,1,0,0),
  oxycodone_po = c(1,NA, 1, 1),
  oxymorphone_po = c(1,2,0,0),
  sufentanil_iv = c(1,2,0,0),
  tramadol_iv = c(1,1,NA,NA)
  )

meq <- meq_po
medd(data)
}
\keyword{medd}

