% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{meq_po}
\alias{meq_po}
\title{Morphine Equivalent to PO}
\format{A data frame with 2 variables:
\describe{
  \item{med_route}{generic name and route}
  \item{meq}{conversion factor to orphine}
  ...
}}
\source{
Needs to be defined
}
\usage{
meq_po
}
\description{
Used to convert various opoid dosages
to Morphine equivalent
}
\keyword{datasets}

